+++ src/util.js
@@ -48,3 +48,21 @@ export function microSeconds() {
         return ms * 1000;
     }
 }
+
+/**
+ * Check if WebLock API is supported.
+ * @link https://developer.mozilla.org/en-US/docs/Web/API/Web_Locks_API
+ */
+export function supportsWebLockAPI() {
+    if (
+        typeof navigator !== 'undefined' &&
+        typeof navigator.locks !== 'undefined' &&
+        typeof navigator.locks.request === 'function'
+    ) {
+        console.log('supportsWebLockAPI(): true');
+        return true;
+    } else {
+        console.log('supportsWebLockAPI(): false');
+        return false;
+    }
+}